+++ src/components/basic_table/collapsed_item_actions.js
@@ -2,6 +2,7 @@ import React, { Component } from 'react';
 import { EuiContextMenuItem, EuiContextMenuPanel } from '../context_menu';
 import { EuiPopover } from '../popover';
 import { EuiButtonIcon } from '../button';
+import { EuiToolTip } from '../tool_tip';
 
 export class CollapsedItemActions extends Component {
 
@@ -89,13 +90,19 @@ export class CollapsedItemActions extends Component {
     const popoverButton = (
       <EuiButtonIcon
         className={className}
-        aria-label="actions"
-        iconType="gear"
+        aria-label="All actions"
+        iconType="boxesHorizontal"
         color="text"
         isDisabled={allDisabled}
         onClick={this.togglePopover.bind(this)}
         onFocus={onFocus}
       />
+    );
+
+    const withTooltip = !allDisabled && (
+      <EuiToolTip content="All actions" delay="long">
+        {popoverButton}
+      </EuiToolTip>
     );
 
     return (
@@ -104,7 +111,7 @@ export class CollapsedItemActions extends Component {
         popoverRef={this.registerPopoverDiv}
         id={`${itemId}-actions`}
         isOpen={isOpen}
-        button={popoverButton}
+        button={withTooltip || popoverButton}
         closePopover={this.closePopover}
         panelPaddingSize="none"
         anchorPosition="leftCenter"