+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 - Fixed issue with unselected tabs and aria-controls attribute in EuiTabbedContent
 
+## [`4.0.0`](https://github.com/elastic/eui/tree/v4.0.0)
+
+- Added `delay` prop to `EuiToolTip` ([#1103](https://github.com/elastic/eui/pull/1103))
+
+**Breaking changes**
+
+- `EuiBasicTable` now shows up to 2 actions before condensing to all popover, but still displaying the top/primary 2 actions as well ([#1103](https://github.com/elastic/eui/pull/1103))
+- `EuiBasicTable` will automatically add `hasActions` and `isSelectable` to allow proper responsive style handling, but are still overridable ([#1103](https://github.com/elastic/eui/pull/1103))
+
 ## [`3.11.0`](https://github.com/elastic/eui/tree/v3.11.0)
 
 - Decorated `pagination` _next_ and _previous_ buttons with `data-test-subj`. ([#1182](https://github.com/elastic/eui/pull/1182))